requirements.txt
# System//Zero Dependencies

# Core
pyyaml>=6.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0  # For FastAPI test client
pytest-cov>=4.1.0

# Development
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
